+++ Src/Private/Get-DiagBackupToHvProxy.ps1
@@ -5,7 +5,7 @@ function Get-DiagBackupToHvProxy {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        0.6.26
+        Version:        0.6.29
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -27,15 +27,15 @@ function Get-DiagBackupToHvProxy {
             $HyperVBackupProxy = Get-VbrBackupProxyInfo -Type 'hyperv'
             if ($HyperVBackupProxy) {
 
-                $columnSize = & {
-                    if (($HyperVBackupProxy | Measure-Object).count -le 1 ) {
-                        return 1
-                    } else {
-                        return 4
-                    }
+                if ($HyperVBackupProxy.Name.Count -eq 1) {
+                    $HyperVBackupProxyColumnSize = 1
+                } elseif ($ColumnSize) {
+                    $HyperVBackupProxyColumnSize = $ColumnSize
+                } else {
+                    $HyperVBackupProxyColumnSize = $HyperVBackupProxy.Name.Count
                 }
 
-                Node HvProxies @{Label = (Get-DiaHTMLNodeTable -ImagesObj $Images -inputObject ($HyperVBackupProxy | ForEach-Object { $_.Name.split('.')[0] }) -Align "Center" -iconType "VBR_Proxy_Server" -columnSize $columnSize -IconDebug $IconDebug -MultiIcon -AditionalInfo $HyperVBackupProxy.AditionalInfo -Subgraph -SubgraphIconType "VBR_HyperV" -SubgraphLabel "Hyper-V Backup Proxies" -SubgraphLabelPos "top" -SubgraphTableStyle "dashed,rounded" -fontColor $Fontcolor -TableBorderColor $Edgecolor -TableBorder "1" -fontSize 18 -SubgraphLabelFontsize 22); shape = 'plain'; fontsize = 18; fontname = "Segoe Ui" }
+                Node HvProxies @{Label = (Get-DiaHTMLNodeTable -ImagesObj $Images -inputObject ($HyperVBackupProxy | ForEach-Object { $_.Name.split('.')[0] }) -Align "Center" -iconType "VBR_Proxy_Server" -columnSize $HyperVBackupProxyColumnSize -IconDebug $IconDebug -MultiIcon -AditionalInfo $HyperVBackupProxy.AditionalInfo -Subgraph -SubgraphIconType "VBR_HyperV" -SubgraphLabel "Hyper-V Backup Proxies" -SubgraphLabelPos "top" -SubgraphTableStyle "dashed,rounded" -fontColor $Fontcolor -TableBorderColor $Edgecolor -TableBorder "1" -fontSize 18 -SubgraphLabelFontsize 22); shape = 'plain'; fontsize = 18; fontname = "Segoe Ui" }
 
                 Edge BackupServers -To HvProxies @{minlen = 3 }
             }
@@ -48,15 +48,32 @@ function Get-DiagBackupToHvProxy {
                     $vCenterNodeArray = @()
                     $ViClustersNodes = @()
                     $vCenterNodeArray += $vCenter.Label
+
                     try {
-                        $ViClustersChildsNodes = Get-DiaHTMLTable -ImagesObj $Images -Rows $vCenter.Childs.Name -Align 'Center' -ColumnSize 3 -IconDebug $IconDebug -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -NoFontBold -fontSize 18
+                        if ($vCenter.Childs.Name.Count -eq 1) {
+                            $HyperVBackupProxyColumnSize = 1
+                        } elseif ($ColumnSize) {
+                            $HyperVBackupProxyColumnSize = $ColumnSize
+                        } else {
+                            $HyperVBackupProxyColumnSize = $vCenter.Childs.Name.Count
+                        }
+
+                        $ViClustersChildsNodes = Get-DiaHTMLTable -ImagesObj $Images -Rows $vCenter.Childs.Name -Align 'Center' -ColumnSize $HyperVBackupProxyColumnSize -IconDebug $IconDebug -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -NoFontBold -FontSize 18
+
                     } catch {
                         Write-Verbose "Error: Unable to create Hyper-V Hosts table Objects. Disabling the section"
                         Write-Debug "Error Message: $($_.Exception.Message)"
                     }
                     try {
                         if ($ViClustersChildsNodes) {
-                            $ViClustersNodes += Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ViClustersChildsNodes -Align 'Center' -IconDebug $IconDebug -Label 'Hosts' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 3 -fontSize 18
+                            if ($ViClustersChildsNodes.Count -eq 1) {
+                                $ViClustersNodesColumnSize = 1
+                            } elseif ($ColumnSize) {
+                                $ViClustersNodesColumnSize = $ColumnSize
+                            } else {
+                                $ViClustersNodesColumnSize = $ViClustersChildsNodes.Count
+                            }
+                            $ViClustersNodes += Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ViClustersChildsNodes -Align 'Center' -IconDebug $IconDebug -Label 'Hosts' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize $ViClustersNodesColumnSize -fontSize 18
                             $vCenterNodeArray += $ViClustersNodes
                         }
                     } catch {
@@ -75,7 +92,14 @@ function Get-DiagBackupToHvProxy {
 
                 try {
                     if ($vCenterNodeArray) {
-                        $VivCenterNodesAll += Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $VivCenterNodes -Align 'Center' -IconDebug $IconDebug -Label 'Hyper-V Clusters' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 3 -fontSize 22
+                        if ($VivCenterNodes.Count -eq 1) {
+                            $VivCenterNodesAllColumnSize = 1
+                        } elseif ($ColumnSize) {
+                            $VivCenterNodesAllColumnSize = $ColumnSize
+                        } else {
+                            $VivCenterNodesAllColumnSize = $VivCenterNodes.Count
+                        }
+                        $VivCenterNodesAll += Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $VivCenterNodes -Align 'Center' -IconDebug $IconDebug -Label 'Hyper-V Clusters' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize $VivCenterNodesAllColumnSize -fontSize 22
                     }
                 } catch {
                     Write-Verbose "Error: Unable to create Hyper-V Cluster Objects. Disabling the section"
@@ -85,17 +109,17 @@ function Get-DiagBackupToHvProxy {
 
             if ($HyperVServerObj = Get-VbrBackupHyperVStandAloneInfo | Sort-Object) {
 
-                $columnSize = & {
-                    if (($HyperVServerObj | Measure-Object).count -le 1 ) {
-                        return 1
-                    } else {
-                        return 4
-                    }
+                if ($HyperVServerObj.Name.Count -eq 1) {
+                    $HyperVServerObjColumnSize = 1
+                } elseif ($ColumnSize) {
+                    $HyperVServerObjColumnSize = $ColumnSize
+                } else {
+                    $HyperVServerObjColumnSize = $HyperVServerObj.Name.Count
                 }
 
                 try {
 
-                    $ViStandAloneNodes = (Get-DiaHTMLNodeTable -ImagesObj $Images -inputObject ($HyperVServerObj | ForEach-Object { $_.Name.split('.')[0] }) -Align "Center" -iconType "VBR_HyperV_Server" -columnSize $columnSize -IconDebug $IconDebug -MultiIcon -AditionalInfo $HyperVServerObj.AditionalInfo -Subgraph -SubgraphLabel " " -SubgraphLabelPos "top" -SubgraphTableStyle "dashed,rounded" -fontColor $Fontcolor -TableBorderColor $Edgecolor -TableBorder "1")
+                    $ViStandAloneNodes = (Get-DiaHTMLNodeTable -ImagesObj $Images -inputObject ($HyperVServerObj | ForEach-Object { $_.Name.split('.')[0] }) -Align "Center" -iconType "VBR_HyperV_Server" -columnSize $HyperVServerObjColumnSize -IconDebug $IconDebug -MultiIcon -AditionalInfo $HyperVServerObj.AditionalInfo -Subgraph -SubgraphLabel " " -SubgraphLabelPos "top" -SubgraphTableStyle "dashed,rounded" -fontColor $Fontcolor -TableBorderColor $Edgecolor -TableBorder "1")
                 } catch {
                     Write-Verbose "Error: Unable to create Hyper-V StandAlone Hosts Table. Disabling the section"
                     Write-Debug "Error Message: $($_.Exception.Message)"
@@ -103,7 +127,7 @@ function Get-DiagBackupToHvProxy {
 
                 if ($ViStandAloneNodes) {
                     try {
-                        $VivCenterNodesAll += Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ViStandAloneNodes -Align 'Center' -IconDebug $IconDebug -Label 'Hyper-V StandAlone Hosts' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize $columnSize -fontSize 22
+                        $VivCenterNodesAll += Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ViStandAloneNodes -Align 'Center' -IconDebug $IconDebug -Label 'Hyper-V StandAlone Hosts' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 1 -fontSize 22
                     } catch {
                         Write-Verbose "Error: Unable to create Hyper-V StandAlone Objects. Disabling the section"
                         Write-Debug "Error Message: $($_.Exception.Message)"
@@ -112,6 +136,7 @@ function Get-DiagBackupToHvProxy {
             }
 
             if ($VivCenterNodesAll) {
+
                 if ($Dir -eq 'LR') {
                     try {
                         $ViClustersSubgraphNode = Node -Name "HvCluster" -Attributes @{Label = (Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $VivCenterNodesAll -Align 'Center' -IconDebug $IconDebug -IconType 'VBR_HyperV' -Label 'Microsoft Hyper-V Infrastructure' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 1 -fontSize 24); shape = 'plain'; fillColor = 'transparent'; fontsize = 14; fontname = "Segoe Ui" }