+++ Src/Private/Get-DiagBackupToProtectedGroup.ps1
@@ -5,7 +5,7 @@ function Get-DiagBackupToProtectedGroup {
     .DESCRIPTION
         Build a diagram of the configuration of Veeam VBR in PDF/PNG/SVG formats using Psgraph.
     .NOTES
-        Version:        0.6.19
+        Version:        0.6.29
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -30,22 +30,28 @@ function Get-DiagBackupToProtectedGroup {
             $IndividualContainer = $ProtectedGroups | Where-Object { $_.Container -eq 'IndividualComputers' }
             $CSVContainer = $ProtectedGroups | Where-Object { $_.Container -eq 'CSV' }
 
-            try {
-                $FileBackupProxy = Get-VbrBackupProxyInfo -Type 'nas'
-                if ($BackupServerInfo) {
-                    if ($FileBackupProxy) {
-
-                        Node FileProxies @{Label = (Get-DiaHTMLNodeTable -ImagesObj $Images -inputObject ($FileBackupProxy | ForEach-Object { $_.Name.split('.')[0] }) -Align "Center" -iconType "VBR_Proxy_Server" -columnSize 4 -IconDebug $IconDebug -MultiIcon -AditionalInfo $FileBackupProxy.AditionalInfo -Subgraph -SubgraphIconType "VBR_Proxy" -SubgraphLabel "File Backup Proxies" -SubgraphLabelPos "top" -SubgraphTableStyle "dashed,rounded" -fontColor $Fontcolor -TableBorderColor $Edgecolor -TableBorder "1" -SubgraphLabelFontsize 24 -fontSize 18); shape = 'plain'; fontsize = 14; fontname = "Segoe Ui" }
-
-                        Edge BackupServers -To FileProxies @{minlen = 3 }
-
-                    }
-                }
-            } catch {
-                Write-Verbose -Message $_.Exception.Message
-            }
-
             if ($ProtectedGroups.Container) {
+                try {
+                    $FileBackupProxy = Get-VbrBackupProxyInfo -Type 'nas'
+                    if ($BackupServerInfo) {
+                        if ($FileBackupProxy) {
+                            if ($FileBackupProxy.Name.Count -eq 1) {
+                                $FileBackupProxyColumnSize = 1
+                            } elseif ($ColumnSize) {
+                                $FileBackupProxyColumnSize = $ColumnSize
+                            } else {
+                                $FileBackupProxyColumnSize = $FileBackupProxy.Name.Count
+                            }
+
+                            Node FileProxies @{Label = (Get-DiaHTMLNodeTable -ImagesObj $Images -inputObject ($FileBackupProxy | ForEach-Object { $_.Name.split('.')[0] }) -Align "Center" -iconType "VBR_Proxy_Server" -columnSize $FileBackupProxyColumnSize -IconDebug $IconDebug -MultiIcon -AditionalInfo $FileBackupProxy.AditionalInfo -Subgraph -SubgraphIconType "VBR_Proxy" -SubgraphLabel "File Backup Proxies" -SubgraphLabelPos "top" -SubgraphTableStyle "dashed,rounded" -fontColor $Fontcolor -TableBorderColor $Edgecolor -TableBorder "1" -SubgraphLabelFontsize 24 -fontSize 18); shape = 'plain'; fontsize = 14; fontname = "Segoe Ui" }
+
+                            Edge BackupServers -To FileProxies @{minlen = 3 }
+
+                        }
+                    }
+                } catch {
+                    Write-Verbose -Message $_.Exception.Message
+                }
                 if ($ProtectedGroups) {
                     $ComputerAgentsArray = @()
                     if ($ADContainer) {
@@ -79,8 +85,15 @@ function Get-DiagBackupToProtectedGroup {
                         }
 
                         if ($ADCNodes) {
+                            if ($ADCNodes.Count -eq 1) {
+                                $ADCNodesColumnSize = 1
+                            } elseif ($ColumnSize) {
+                                $ADCNodesColumnSize = $ColumnSize
+                            } else {
+                                $ADCNodesColumnSize = $ADCNodes.Count
+                            }
                             try {
-                                $ADCNodesSubgraph = Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ADCNodes -Align 'Center' -IconDebug $IconDebug -Label 'Active Directory Computers' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 2 -IconType "VBR_AGENT_AD" -fontSize 18
+                                $ADCNodesSubgraph = Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ADCNodes -Align 'Center' -IconDebug $IconDebug -Label 'Active Directory Computers' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize $ADCNodesColumnSize -IconType "VBR_AGENT_AD" -fontSize 18
                             } catch {
                                 Write-Verbose "Error: Unable to create ADCNodesSubgraph Objects. Disabling the section"
                                 Write-Debug "Error Message: $($_.Exception.Message)"
@@ -112,8 +125,15 @@ function Get-DiagBackupToProtectedGroup {
                         }
 
                         if ($MCNodes) {
+                            if ($MCNodes.Count -eq 1) {
+                                $MCNodesColumnSize = 1
+                            } elseif ($ColumnSize) {
+                                $MCNodesColumnSize = $ColumnSize
+                            } else {
+                                $MCNodesColumnSize = $MCNodes.Count
+                            }
                             try {
-                                $MCNodesSubgraph = Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $MCNodes -Align 'Center' -IconDebug $IconDebug -Label 'Manual Computers' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 2 -IconType "VBR_AGENT_MC" -fontSize 18
+                                $MCNodesSubgraph = Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $MCNodes -Align 'Center' -IconDebug $IconDebug -Label 'Manual Computers' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize $MCNodesColumnSize -IconType "VBR_AGENT_MC" -fontSize 18
                             } catch {
                                 Write-Verbose "Error: Unable to create MCNodesSubgraph Objects. Disabling the section"
                                 Write-Debug "Error Message: $($_.Exception.Message)"
@@ -153,8 +173,15 @@ function Get-DiagBackupToProtectedGroup {
                         }
 
                         if ($ICCNodes) {
+                            if ($ICCNodes.Count -eq 1) {
+                                $ICCNodesColumnSize = 1
+                            } elseif ($ColumnSize) {
+                                $ICCNodesColumnSize = $ColumnSize
+                            } else {
+                                $ICCNodesColumnSize = $ICCNodes.Count
+                            }
                             try {
-                                $ICCNodesSubgraph = Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ICCNodes -Align 'Center' -IconDebug $IconDebug -Label 'Individual Computers' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 2 -IconType "VBR_AGENT_IC" -fontSize 18
+                                $ICCNodesSubgraph = Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ICCNodes -Align 'Center' -IconDebug $IconDebug -Label 'Individual Computers' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize $ICCNodesColumnSize -IconType "VBR_AGENT_IC" -fontSize 18
                             } catch {
                                 Write-Verbose "Error: Unable to create ICCNodesSubgraph Objects. Disabling the section"
                                 Write-Debug "Error Message: $($_.Exception.Message)"
@@ -182,8 +209,15 @@ function Get-DiagBackupToProtectedGroup {
                         }
 
                         if ($CSVCNodes) {
+                            if ($CSVCNodes.Count -eq 1) {
+                                $CSVCNodesColumnSize = 1
+                            } elseif ($ColumnSize) {
+                                $CSVCNodesColumnSize = $ColumnSize
+                            } else {
+                                $CSVCNodesColumnSize = $CSVCNodes.Count
+                            }
                             try {
-                                $CSVCNodesSubgraph = Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $CSVCNodes -Align 'Center' -IconDebug $IconDebug -Label 'CSV Computers' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 2 -IconType "VBR_AGENT_CSV_Logo" -fontSize 18
+                                $CSVCNodesSubgraph = Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $CSVCNodes -Align 'Center' -IconDebug $IconDebug -Label 'CSV Computers' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize $CSVCNodesColumnSize -IconType "VBR_AGENT_CSV_Logo" -fontSize 18
                             } catch {
                                 Write-Verbose "Error: Unable to create CSVCNodesSubgraph Objects. Disabling the section"
                                 Write-Debug "Error Message: $($_.Exception.Message)"
@@ -193,16 +227,23 @@ function Get-DiagBackupToProtectedGroup {
                     }
 
                     if ($ComputerAgentsArray) {
+                        if ($ComputerAgentsArray.Count -eq 1) {
+                            $ComputerAgentsArrayColumnSize = 1
+                        } elseif ($ColumnSize) {
+                            $ComputerAgentsArrayColumnSize = $ColumnSize
+                        } else {
+                            $ComputerAgentsArrayColumnSize = $ComputerAgentsArray.Count
+                        }
                         if ($Dir -eq 'LR') {
                             try {
-                                $ComputerAgentSubGraph = Node -Name "ComputerAgentsSubgraph" -Attributes @{Label = (Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ComputerAgentsArray -Align 'Center' -IconDebug $IconDebug -Label 'Protected Groups' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 1 -fontSize 26); shape = 'plain'; fillColor = 'transparent'; fontsize = 14; fontname = "Segoe Ui" }
+                                $ComputerAgentSubGraph = Node -Name "ComputerAgentsSubgraph" -Attributes @{Label = (Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ComputerAgentsArray -Align 'Center' -IconDebug $IconDebug -Label 'Protected Groups' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize $ComputerAgentsArrayColumnSize -fontSize 26); shape = 'plain'; fillColor = 'transparent'; fontsize = 14; fontname = "Segoe Ui" }
                             } catch {
                                 Write-Verbose "Error: Unable to create ComputerAgentsSubgraph Objects. Disabling the section"
                                 Write-Debug "Error Message: $($_.Exception.Message)"
                             }
                         } else {
                             try {
-                                $ComputerAgentSubGraph = Node -Name "ComputerAgentsSubgraph" -Attributes @{Label = (Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ComputerAgentsArray -Align 'Center' -IconDebug $IconDebug -Label 'Protected Groups' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize 4 -fontSize 26); shape = 'plain'; fillColor = 'transparent'; fontsize = 14; fontname = "Segoe Ui" }
+                                $ComputerAgentSubGraph = Node -Name "ComputerAgentsSubgraph" -Attributes @{Label = (Get-DiaHTMLSubGraph -ImagesObj $Images -TableArray $ComputerAgentsArray -Align 'Center' -IconDebug $IconDebug -Label 'Protected Groups' -LabelPos "top" -fontColor $Fontcolor -TableStyle "dashed,rounded" -TableBorderColor $Edgecolor -TableBorder "1" -columnSize $ComputerAgentsArrayColumnSize -fontSize 26); shape = 'plain'; fillColor = 'transparent'; fontsize = 14; fontname = "Segoe Ui" }
                             } catch {
                                 Write-Verbose "Error: Unable to create ComputerAgentsSubgraph Objects. Disabling the section"
                                 Write-Debug "Error Message: $($_.Exception.Message)"